splitting components
