+++ src/stripe.js
@@ -1,93 +1,10 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useState } from 'react';
+import React from 'react';
 // import axios from 'axios';
 import { loadStripe } from '@stripe/stripe-js';
-import {
-  Elements,
-  CardElement,
-  useStripe,
-  useElements,
-} from '@stripe/react-stripe-js';
-
-const CARD_ELEMENT_OPTIONS = {
-  style: {
-    base: {
-      color: '#32325d',
-      fontFamily: '"Helvetica Neue", Helvetica, sans-serif',
-      fontSmoothing: 'antialiased',
-      fontSize: '16px',
-      '::placeholder': {
-        color: '#aab7c4',
-      },
-    },
-    invalid: {
-      color: '#fa755a',
-      iconColor: '#fa755a',
-    },
-  },
-};
-// POST el id del token hacia el backend:
-async function stripeTokenHandler(token) {
-  const response = await fetch('Api/charge', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ token: token.id }),
-  });
-
-  return response.json();
-}
-
-const CheckoutForm = () => {
-  const stripe = useStripe();
-  const elements = useElements();
-  const [error, setError] = useState(null);
-
-  const handleChange = (event) => {
-    if (event.error) {
-      setError(event.error.message);
-    } else {
-      setError(null);
-    }
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const card = elements.getElement(CardElement);
-    const result = await stripe.createToken(card);
-    if (result.error) {
-      setError(result.error.message);
-    } else {
-      setError(null);
-      stripeTokenHandler(result.token);
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <div className="form-row">
-        <label htmlFor="card-element">
-          Credit or debit card
-        </label>
-        <CardElement
-          id="card-element"
-          onChange={handleChange}
-          options={CARD_ELEMENT_OPTIONS}
-        />
-        <div className="card-errors" role="alert">{error}</div>
-      </div>
-      <button type="submit">Submit Payment</button>
-    </form>
-    // <form onSubmit={handleSubmit}>
-    //   <h2>Product name</h2>
-    //   <h2>$150.00</h2>
-    //   <CardElement />
-    //   <button type="submit" disabled={!stripe}>Pay</button>
-    // </form>
-  );
-};
+import { Elements } from '@stripe/react-stripe-js';
+import CheckoutForm from './containers/CheckoutForm';
 
 const stripePromise = loadStripe('pk_test_s9yfSxszCqB1k7qTbuUMF4wQ00uWkgY9ZC');
